Add a new BaseTest class
It provides a clear and documented way to write additional test classes.

It also features a new output streaming capability: it is now possible to
dynamically print the output of a test as it runs (to stdout or a random file)
while still keeping a copy in the object for later uses.
Please note that the output is now combined: there is no longer 'stdout' and
'stderr', only 'output'.

LocalTest and RemoteTest have to be rewritten to take advantage of it.