Add the --remote option to define remotes that override those from lift.yaml
This can be useful to quickly run a test on another remote: it is no longer
necessary to edit the remote definition in the lift.yaml file.

This is also interesting in the context of automated runs where remotes
addresses can not be predicted in advance. It is easier to generate a command
line defining the right remotes than to generate a whole lift.yaml file.